Migrate to composable pattern
